# api_server/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is Vacarion Calendar API description"
  version: "1.0.0"
  title: "Swagger Vacation calendar"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "w.kostyukova@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/vcalendar"
...
schemes:
- "https"
- "http"
paths:
...
  /employee/total_days:
    post:
      tags:
      - "total_days"
      summary: "Add a employee total days to the system. Role write:total_days must\
        \ be granted"
      description: ""
      operationId: "add_total_days"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "TotalDays object that needs to be added to the system"
        required: true
        schema:
          $ref: "#/definitions/TotalDays"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/TotalDays"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ApiResponse"
        "409":
          description: "TodalDays already exists conflict"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - Bearer: []
      x-swagger-router-controller: "swagger_server.controllers.total_days_controller"
  /employee/total_days/findBy:
    get:
      tags:
      - "total_days"
      summary: "Finds TotalDays by given parameters"
      description: ""
      operationId: "find_total_days_by"
      produces:
      - "application/json"
      parameters:
      - name: "employee_id"
        in: "query"
        description: "Employee id to filter by"
        required: false
        type: "integer"
      - name: "year"
        in: "query"
        description: "Year to filter by"
        required: false
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TotalDays"
        "400":
          description: "Invalid params to filter by"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "swagger_server.controllers.total_days_controller"
  /employee/{employeeId}/total_days/{year}:
    get:
      tags:
      - "total_days"
      summary: "Finds unique TotalDay by employee and year"
      description: ""
      operationId: "find_total_days_by_year"
      produces:
      - "application/json"
      parameters:
      - name: "employeeId"
        in: "path"
        description: "Employee id to filter by"
        required: true
        type: "integer"
      - name: "year"
        in: "path"
        description: "Year to filter by"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/TotalDays"
        "400":
          description: "Invalid name supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "TotalDays not found"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "swagger_server.controllers.total_days_controller"
  /employee/{employeeId}/total_days/:
    get:
      tags:
      - "total_days"
      summary: "Finds TotalDays by employee"
      description: ""
      operationId: "find_employee_total_days"
      produces:
      - "application/json"
      parameters:
      - name: "employeeId"
        in: "path"
        description: "Employee id to filter by"
        required: true
        type: "integer"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TotalDays"
        "404":
          description: "Employee not found"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "swagger_server.controllers.total_days_controller"
  /employee/total_days/{id}:
    get:
      tags:
      - "total_days"
      summary: "Find total days by ID"
      description: "Returns a single TotalDays."
      operationId: "get_total_days_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of TotalDays to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/TotalDays"
        "400":
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "TotalDays not found"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "swagger_server.controllers.total_days_controller"
    put:
      tags:
      - "total_days"
      summary: "Updates a TotalDays in the system with form data. Role write:total_days\
        \ must be granted"
      description: ""
      operationId: "update_total_days_by_id"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of TotalDays to return"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "TotalDays object that needs to be updated in the system"
        required: true
        schema:
          $ref: "#/definitions/TotalDays"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/TotalDays"
        "400":
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "TotalDays not found"
          schema:
            $ref: "#/definitions/ApiResponse"
        "405":
          description: "Invalid input"
          schema:
            $ref: "#/definitions/ApiResponse"
        "409":
          description: "TotalDays already exists conflict"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - Bearer: []
      x-swagger-router-controller: "swagger_server.controllers.total_days_controller"
    delete:
      tags:
      - "total_days"
      summary: "Deletes a TotalDays. Role write:total_days must be granteds"
      description: ""
      operationId: "delete_total_days"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "TotalDays id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "204":
          description: "Successful deletion"
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "TotalDays not found"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - Bearer: []
      x-swagger-router-controller: "swagger_server.controllers.total_days_controller"
...
  /user:
    post:
      tags:
      - "user"
      summary: "Create user"
      description: "Role write:users role must be granted"
      operationId: "create_user"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/User"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ApiResponse"
        "409":
          description: "User already exists conflict"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - Bearer: []
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /user/authenticate:
    get:
      tags:
      - "user"
      summary: "Generates token for an user"
      description: ""
      operationId: "authenticate_user"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "query"
        description: "The user name for login"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "The password for login in clear text"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "string"
        "400":
          description: "Invalid username/password supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /user/:
    get:
      tags:
      - "user"
      summary: "Gets user by user name. Role read:users role must be granted"
      description: ""
      operationId: "get_user_by_name"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "query"
        description: "The name of user that needs to be fetched. "
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Invalid username supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - Bearer: []
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /user/{id}:
    get:
      tags:
      - "user"
      summary: "Gets user by id. Role read:users role must be granted"
      description: ""
      operationId: "get_user_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of user that needs to be fetched"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Invalid username supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - Bearer: []
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
    put:
      tags:
      - "user"
      summary: "Updated user"
      description: "This can only be done by with write:users role."
      operationId: "update_user"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of user that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Invalid user supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - Bearer: []
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
    delete:
      tags:
      - "user"
      summary: "Delete user"
      description: "This can only be done with write:users role."
      operationId: "delete_user"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of user that needs to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "204":
          description: "Successful deletion"
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: "Invalid username supplied"
          schema:
            $ref: "#/definitions/ApiResponse"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ApiResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - Bearer: []
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /user/findBy:
    get:
      tags:
      - "user"
      summary: "Finds users by given parameters. Role read:users role must be granted"
      description: ""
      operationId: "find_by"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "query"
        description: "The username to filter by. "
        required: false
        type: "string"
      - name: "email"
        in: "query"
        description: "The email to filter by. "
        required: false
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        "403":
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - Bearer: []
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
        example: "user"
      email:
        type: "string"
        example: "user@mail.com"
      password:
        type: "string"
        example: "123qwe"
      roles:
        type: "string"
        example: "write:employees"
    xml:
      name: "User"
    example:
      password: "123qwe"
      roles: "write:employees"
      id: 0
      email: "user@mail.com"
      username: "user"
...
  TotalDays:
    type: "object"
    required:
    - "employee_id"
    - "total_days"
    - "year"
    properties:
      id:
        type: "integer"
        format: "int64"
      employee_id:
        type: "integer"
        format: "int64"
      total_days:
        type: "integer"
        format: "int32"
        example: 20
      year:
        type: "integer"
        format: "int32"
        example: 2020
    example:
      year: 2020
      employee_id: 6
      id: 0
      total_days: 20
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
    example:
      code: 0
      type: "type"
      message: "message"
...